# Lectures/Lecture_11/CookBook/src/app/api/swagger.json
{
  "x-generator": "NSwag v13.2.2.0 (NJsonSchema v10.1.4.0 (Newtonsoft.Json v12.0.0.0))",
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:44378"
    }
  ],
  "paths": {
    "/api/Ingredient": {
      "get": {
        "tags": [
          "Ingredient"
        ],
        "operationId": "IngredientGetAll",
        "parameters": [
          {
            "name": "version",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "type": "string",
            "name": "culture",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IngredientListModel"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Ingredient"
        ],
        "operationId": "IngredientCreate",
        "parameters": [
          {
            "name": "version",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "type": "string",
            "name": "culture",
            "in": "query"
          }
        ],
        "requestBody": {
          "x-name": "ingredient",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientDetailModel"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "guid"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Ingredient"
        ],
        "operationId": "IngredientUpdate",
        "parameters": [
          {
            "name": "version",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "type": "string",
            "name": "culture",
            "in": "query"
          }
        ],
        "requestBody": {
          "x-name": "ingredient",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientDetailModel"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "guid"
                }
              }
            }
          }
        }
      }
    },
    "/api/Ingredient/{id}": {
      "get": {
        "tags": [
          "Ingredient"
        ],
        "operationId": "IngredientGetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "version",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "type": "string",
            "name": "culture",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientDetailModel"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Ingredient"
        ],
        "operationId": "IngredientDelete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "version",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "type": "string",
            "name": "culture",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Logging": {
      "get": {
        "tags": [
          "Logging"
        ],
        "operationId": "LoggingLogSomething",
        "parameters": [
          {
            "name": "version",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "type": "string",
            "name": "culture",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/Recipe": {
      "get": {
        "tags": [
          "Recipe"
        ],
        "operationId": "RecipeGetAll",
        "parameters": [
          {
            "name": "version",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "type": "string",
            "name": "culture",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecipeListModel"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Recipe"
        ],
        "operationId": "RecipeCreate",
        "parameters": [
          {
            "name": "version",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "type": "string",
            "name": "culture",
            "in": "query"
          }
        ],
        "requestBody": {
          "x-name": "recipe",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeDetailModel"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "guid"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Recipe"
        ],
        "operationId": "RecipeUpdate",
        "parameters": [
          {
            "name": "version",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "type": "string",
            "name": "culture",
            "in": "query"
          }
        ],
        "requestBody": {
          "x-name": "recipe",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeDetailModel"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "guid",
                  "nullable": true
                }
              }
            }
          }
        }
      }
    },
    "/api/Recipe/{id}": {
      "get": {
        "tags": [
          "Recipe"
        ],
        "operationId": "RecipeGetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "version",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "type": "string",
            "name": "culture",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDetailModel"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Recipe"
        ],
        "operationId": "RecipeDelete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "version",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "type": "string",
            "name": "culture",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "IngredientListModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "IngredientDetailModel": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "description"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "minLength": 10
          }
        }
      },
      "RecipeListModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "duration": {
            "type": "string",
            "format": "time-span"
          },
          "foodType": {
            "$ref": "#/components/schemas/FoodType"
          },
          "foodTypeText": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "FoodType": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Unknown",
          "MainDish",
          "Soup",
          "Dessert"
        ],
        "enum": [
          0,
          1,
          2,
          3
        ]
      },
      "RecipeDetailModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "duration": {
            "type": "string",
            "format": "time-span"
          },
          "foodType": {
            "$ref": "#/components/schemas/FoodType"
          },
          "ingredients": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/RecipeListIngredientModel"
            }
          }
        }
      },
      "RecipeListIngredientModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "ingredient": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/IngredientListModel"
              }
            ]
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "unit": {
            "$ref": "#/components/schemas/Unit"
          },
          "unitText": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "Unit": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Unknown",
          "Kg",
          "L",
          "Ml",
          "G",
          "Pieces",
          "Spoon"
        ],
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6
        ]
      }
    }
  }
}